\name{callNameArgs}
\alias{callName}
\title{Get the name or argument list of a call...}
\usage{callName(call)
callArgs(call)
}
\description{Get the name or argument list of a call}
\details{\code{callName}: \code{callName} returns the name of a \code{\link{call}}.
\code{callArgs} returns the argument list of a call.

}
\value{\code{callName}: The name or argument list of a call.

}
\arguments{\item{call}{The call to return the name or argument list for.}
}
\alias{callArgs}

